# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Application API endpoints protected by Better Auth session cookies. Health is liveness only.",
        "title": "AI Interview Practice API",
        "contact": {},
        "version": "1.0.0"
    },
    "basePath": "/",
    "paths": {
        "/auth/me": {
            "get": {
                "security": [
                    {
                        "SessionCookie": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Current application user",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/auth.User"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "apperror.Code": {
            "type": "string",
            "enum": [
                "JD_IN_USE",
                "JD_NOT_FOUND",
                "INVALID_CREDENTIALS",
                "USER_NOT_FOUND",
                "USER_INACTIVE",
                "INVALID_TOKEN",
                "AUTH_UNAVAILABLE",
                "FORBIDDEN",
                "VALIDATION_ERROR",
                "INTERNAL_ERROR",
                "INVALID_JD_INPUT",
                "JD_TOO_SHORT",
                "JD_TOO_LONG",
                "EXTRACTION_FAILED",
                "INVALID_EXTRACTION_OUTPUT"
            ],
            "x-enum-varnames": [
                "CodeJDInUse",
                "CodeJDNotFound",
                "CodeInvalidCredentials",
                "CodeAccountNotFound",
                "CodeAccountLocked",
                "CodeInvalidToken",
                "CodeAuthUnavailable",
                "CodeForbidden",
                "CodeValidation",
                "CodeInternal",
                "CodeInvalidJDInput",
                "CodeJDTooShort",
                "CodeJDTooLong",
                "CodeExtractionFailed",
                "CodeInvalidExtractionOutput"
            ]
        },
        "auth.User": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "email_verified": {
                    "type": "boolean"
                },
                "full_name": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_locked": {
                    "type": "boolean"
                },
                "role": {
                    "type": "string"
                }
            }
        },
...
        "response.Envelope": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/response.ErrorBody"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "response.ErrorBody": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/apperror.Code"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "SessionCookie": {
            "description": "Better Auth HttpOnly session cookie. Production uses __Secure-rolecue-authentication.session_token; browser sends cookies automatically.",
            "type": "apiKey",
            "name": "Cookie",
            "in": "header"
        }
    }
}